removed column slice step at end of populate_offence and moved util tests into own file This aims to make this function more flexible and useable on other datasets. This function is still rather large and could do with a broader refactor, perhaps separating out the returning police_force column and then the populating Crime_description column
